Fix git `psu` alias when given a remote name
Git aliases, when prefixed with `!` to become a shellscript to run, have
the arguments after the alias pasted to the end of the last command of
the script. This was previously erring because `g psu foo` would end
with `exit 1 foo`, which isn't allowed. Instead we can redirect to null
stdout to silent the arguments after they've been used.